refactor: separate files into folders
