+++ src/lib/engine/render/ceil.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { GenerateSettingsType, Settings } from "../settings";
+import { GenerateSettingsType, Settings } from "../configuration/settings";
 import { Ceil } from "./ceil";
 
 function generateSettings(): GenerateSettingsType {
