+++ src/lib/engine/utils/vector.ts
